+++ Elements/src/Geometry/Line.cs
@@ -152,7 +152,14 @@ namespace Elements.Geometry
         /// <returns></returns>
         public override int GetHashCode()
         {
-            return new[] { this.Start, this.End }.GetHashCode();
+            // https://stackoverflow.com/questions/263400/what-is-the-best-algorithm-for-overriding-gethashcode
+            unchecked
+            {
+                int hash = 17;
+                hash = hash * 23 + this.Start.GetHashCode();
+                hash = hash * 23 + this.End.GetHashCode();
+                return hash;
+            }
         }
 
         /// <summary>
@@ -388,7 +395,7 @@ namespace Elements.Geometry
                 results.Add(Start + d * tMin);
             }
 
-            if (Math.Abs(tMax - tMin) > Vector3.EPSILON && 
+            if (Math.Abs(tMax - tMin) > Vector3.EPSILON &&
                 (infinite || (tMax > -Vector3.EPSILON && tMax < 1 + Vector3.EPSILON)))
             {
                 results.Add(Start + d * tMax);